refactor: move types under utils
